+++ src/app/classes/EventsLog/Mappers/LogsMapper.php
@@ -58,7 +58,7 @@ class LogsMapper extends EntityMapperExtensible
             'null' => true,
         ],
         'createdBy' => [
-            'type' => 'int',
+            'type' => 'bigint',
             'reference_table' => UsersModel::TABLE,
             'reference_field' => 'id',
             'reference_primary_key' => 'id',
