+++ src/app/classes/PiecesPHP/UserSystem/ORM/OTPSecretsUsersMapper.php
@@ -37,7 +37,7 @@ class OTPSecretsUsersMapper extends BaseEntityMapper
             'primary_key' => true,
         ],
         'user' => [
-            'type' => 'int',
+            'type' => 'bigint',
             'reference_table' => UsersModel::TABLE,
             'reference_field' => 'id',
             'reference_primary_key' => 'id',
